new introducer: signed extensible dictionary-based messages! refs #466
This introduces new client and server halves to the Introducer (renaming the
old one with a _V1 suffix). Both have fallbacks to accomodate talking to a
different version.

The V2 protocol sends a three-tuple of (serialized announcement dictionary,
signature, pubkey) for each announcement. The V2 server dispatches messages
to subscribers according to the service-name, but does not otherwise examine
the message or check signatures. The V2 receiver's subscription callback will
receive a (serverid, ann_dict) pair. The 'serverid' will be equal to the
pubkey if all of the following are true:

  the originating client is V2, and was told a privkey to use
  the announcement went through a V2 server
  the signature is valid

If not, 'serverid' will be equal to the tubid portion of the announced FURL,
as was the case for V1 receivers.

There is not yet any code in the tahoe Client to provide a privkey: the
tahoe.cfg mechanism for that is still under consideration. Interaction
between traditional (FURL-based) serverids and new (pubkey-based) serverids
is still an open issue.