upload: remove serverid from Tahoe2PeerSelector
This still returns serverids in the output ('preexisting_shares'), but
uses IServer instances internally. It updates the comments and variable
names in happinessutil.py functions that are now getting servers instead
of serverids (but happinessutil is pretty agnostic about the objects it
manipulates, so none of the code actually changed).

test_upload.py was changed to set up servers instead of serverids, and
storage_client.StubServer was changed to be a singleton when used in
copy/deepcopy (to match NativeStorageServer).

happinessutil.merge_servers was changed to use a manual copy of the
servermap (re-using the actual IServer objects), rather than
copy.deepcopy, to avoid exercising the funky copy/deepcopy methods that
I'd like to get rid of eventually.

happinessutil.shares_by_server() is still called (with serverids) by
encoding.py, but because it doesn't look inside the serverid/IServer
objects it gets, that doesn't matter. When the serverid-ectomy wave
finally makes it to encoding.py, the perhaps-confusing mismatch between
what the caller provides and what shares_by_server()'s docstring says it
receives will be resolved.