Changes test_servers_responding to reflect Upload of Happiness behavior
The test relies upon the old algorithm which would distribute shares
in increasing order to the servers. However, Upload of Happiness
doesn't guarantee this behavior so removing share 9 is not necessarily
removing the share from the server that is set to break.

To get test the same behavior the share 8 must be deleted instead.