create-node: rename "--launch-tor" to "--tor-launch"
I don't like the grammatical flow of this, and it doesn't actually put
all the tor-related arguments next to each other in the --help
output (the flags are put in one block, then the parameters in the
next). But it seems slightly more consistent to start all the
tor-related argument names with a "--tor*" prefix.